Extracting data decoding into a separate module.
