+++ driver/pom.xml
@@ -44,6 +44,11 @@
         </dependency>
         <dependency>
             <groupId>za.co.absa.hyperdrive</groupId>
+            <artifactId>decoder</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>za.co.absa.hyperdrive</groupId>
             <artifactId>writer</artifactId>
             <version>${project.version}</version>
         </dependency>
