+++ driver/src/test/scala/za/co/absa/hyperdrive/driver/TestSparkIngestor.scala
@@ -25,12 +25,12 @@ import org.apache.spark.sql.streaming.{DataStreamReader, StreamingQuery}
 import org.scalatest.{BeforeAndAfterEach, FlatSpec}
 import za.co.absa.hyperdrive.manager.offset.OffsetManager
 import za.co.absa.hyperdrive.reader.StreamReader
-import za.co.absa.hyperdrive.transformer.data.StreamTransformer
 import za.co.absa.hyperdrive.writer.StreamWriter
 import org.scalatest.mockito.MockitoSugar
 import org.mockito.Mockito._
+import za.co.absa.hyperdrive.decoder.StreamDecoder
 import za.co.absa.hyperdrive.shared.exceptions.{IngestionException, IngestionStartException}
-import za.co.absa.hyperdrive.transformer.encoding.StreamDecoder
+import za.co.absa.hyperdrive.transformer.StreamTransformer
 
 class TestSparkIngestor extends FlatSpec with BeforeAndAfterEach with MockitoSugar {
