--- transformer/src/main/scala/za/co/absa/hyperdrive/transformer/data/impl/select/ColumnSelectorStreamTransformer.scala
@@ -1,36 +1,0 @@
-/*
- *  Copyright 2019 ABSA Group Limited
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package za.co.absa.hyperdrive.transformer.data.impl.select
-
-import org.apache.spark.sql.DataFrame
-import za.co.absa.hyperdrive.transformer.data.StreamTransformer
-
-private[data] class ColumnSelectorStreamTransformer(val columns: Seq[String]) extends StreamTransformer {
-
-  if (columns == null || columns.isEmpty) {
-    throw new IllegalArgumentException("Empty list of columns to select.")
-  }
-
-  def transform(streamData: DataFrame): DataFrame = {
-    if (streamData == null) {
-      throw new IllegalArgumentException("Null DataFrame received.")
-    }
-
-    streamData.select(columns.head, columns.tail:_*)
-  }
-}
