+++ trigger/src/main/scala/za/co/absa/hyperdrive/trigger/mock/NotificationDispatcher.scala
@@ -20,9 +20,65 @@ package za.co.absa.hyperdrive.trigger.mock
 
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import za.co.absa.abris.avro.read.confluent.SchemaManager
-import za.co.absa.hyperdrive.shared.InfrastructureSettings.{AvroSettings, HyperdriveSettings, KafkaSettings, SchemaRegistrySettings}
 import za.co.absa.hyperdrive.trigger.notification.Notification
 import ConfigParams._
+
+object SystemSettings {
+  val JVM_DEFAULT_TMP_DIR_PROPERTY = "java.io.tmpdir"
+}
+
+object HyperdriveSettings {
+
+  val BASE_TEMP_DIR: String = {
+    val jvmTmpDir = System.getProperty(SystemSettings.JVM_DEFAULT_TMP_DIR_PROPERTY)
+    val baseTmpDir = if (jvmTmpDir == null || jvmTmpDir.trim.isEmpty) "/tmp" else jvmTmpDir
+    baseTmpDir + "/HYPERDRIVE_TEMP/"
+  }
+  val BASE_PAYLOAD_DESTINATION = "/tmp/HYPERDRIVE_PAYLOAD"
+
+  val NOTIFICATION_TOPIC = "notification_topic"
+  val PAYLOAD_TOPIC_1    = "payload_topic_1"
+  val PAYLOAD_TOPIC_2    = "payload_topic_2"
+  var PAYLOAD_TOPIC_IN_USE: String = PAYLOAD_TOPIC_1
+
+  def getPayloadDestinationDir: String = {
+    if (PAYLOAD_TOPIC_IN_USE == PAYLOAD_TOPIC_1)
+      s"$BASE_PAYLOAD_DESTINATION/dest1"
+    else if (PAYLOAD_TOPIC_IN_USE == PAYLOAD_TOPIC_2)
+      s"$BASE_PAYLOAD_DESTINATION/dest2"
+    else s"$BASE_PAYLOAD_DESTINATION/INCORRECT_DESTINATION"
+  }
+}
+
+object KafkaSettings {
+  val STREAM_FORMAT_KAFKA_NAME  = "kafka"
+  val BROKERS_SETTING_KEY       = "bootstrap.servers"
+  val SPARK_BROKERS_SETTING_KEY = "kafka.bootstrap.servers"
+
+  val TOPIC_SUBSCRIPTION_KEY    = "subscribe"
+  val TOPIC_DISPATCH_KEY        = "topic"
+
+  val STARTING_OFFSETS_KEY      = "startingOffsets"
+  val STARTING_OFFSETS_EARLIEST = "earliest"
+
+  val KEY_DESERIALIZER_KEY      = "key.deserializer"
+  val KEY_DESERIALIZER          = "org.apache.kafka.common.serialization.StringDeserializer"
+
+  val VALUE_DESERIALIZER_KEY    = "value.deserializer"
+  val VALUE_DESERIALIZER        = "org.apache.kafka.common.serialization.ByteArrayDeserializer"
+
+  val GROUP_ID_KEY              = "group.id"
+}
+
+object SchemaRegistrySettings {
+  val KEY_SCHEMA_NAMING_STRATEGY: String   = SchemaManager.SchemaStorageNamingStrategies.TOPIC_NAME
+  val VALUE_SCHEMA_NAMING_STRATEGY: String = SchemaManager.SchemaStorageNamingStrategies.TOPIC_NAME
+}
+
+object AvroSettings {
+  val GENERAL_SCHEMA_NAME      = "schemaName"
+  val GENERAL_SCHEMA_NAMESPACE = "schemaNamespace"
+}
 
 object NotificationDispatcher {
